delete done

+++ resources/js/components/users/Users.vue
@@ -261,7 +261,19 @@ export default {
 
 
         deleteUser(user) {
-           this.$store.dispatch('deleteUser', user)
+            Swal.fire({
+                    title: "Are you sure?",
+                    text: "You won't be able to revert this!",
+                    icon: "warning",
+                    showCancelButton: true,
+                    confirmButtonColor: "#3085d6",
+                    cancelButtonColor: "#d33",
+                    confirmButtonText: "Yes, delete it!"
+                    }).then((result) => {
+                    if (result.isConfirmed) {
+                        this.$store.dispatch('deleteUser', user)
+                    }
+                    });
         },
